Now Wired can be compiled on win32. Added win32 related stuff, fixed some compiler warnings, some code unsupported on win32 were put between pre-compiling directives and some files were renamed to prevent object files with the same name.
git-svn-id: https://svn.code.sf.net/p/wired/code/trunk@1408 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ wired/src/gui/Pattern.cpp
@@ -226,7 +226,7 @@ void					Pattern::OnMotion(wxMouseEvent &e)
 	{
 	  if (SeqPanel->SelectedItems.size() == 1)
 	    {
-	      trackto = TrackIndex + floor((e.GetPosition().y ) / (TRACK_HEIGHT * SeqPanel->VertZoomFactor));
+	      trackto = TrackIndex + (int)floor((e.GetPosition().y ) / (TRACK_HEIGHT * SeqPanel->VertZoomFactor));
 	      if (SeqPanel->IsAudioTrack(TrackIndex) == SeqPanel->IsAudioTrack(trackto))
 		{
 		  SeqPanel->ChangeMouseCursor(wxCursor(wxCURSOR_HAND));
